Refactoring of functions module
